rework value_objects FuncCallPatcherData.pk
Теперь у FuncCallPatcherData появился pk, автоприсваение pk перенес из register в конструктор FuncCallPatcherData
